remove c++ style comments
use %lu and cast off_t to unsigned long in printf()

+++ sync.c
@@ -120,7 +120,7 @@ sync_mailbox (mailbox_t * mbox, imap_t * imap, int flags,
 		    if ((flags & SYNC_QUIET) == 0)
 			printf
 			    ("Warning, local message is too large (%lu), skipping...\n",
-			     sb.st_size);
+			     (unsigned long) sb.st_size);
 		    continue;
 		}
 		fd = open (path, O_RDONLY);
